product_app_backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
black==24.1.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==3.2.16
django-bootstrap4==23.2
django-bootstrap5==23.3
django-braces==1.15.0
django-cors-headers==4.4.0
django-filter==23.5
django-model-utils==4.3.1
django-rest-framework==0.1.0
django-star-ratings==0.9.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.7
filelock==3.13.1
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==21.2.0
identify==2.5.33
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
pep8-naming==0.13.3
Pillow==9.3.0
platformdirs==4.2.0
pluggy==0.13.1
pre-commit==3.6.0
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
swapper==1.3.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.25.0
